requirements.txt
requests~=2.25.1
diffiehellman~=0.13.3
google~=3.0.0
websocket-client~=0.58.0
pycryptodome==3.10.1